ELK for Raspberry Pi
This repo is intended to run an ELK stack on a raspberry pi. The
services seem to work, although I'm hitting an issue: When kibana
initially starts, it creates an index named .kibana. The API request to
create this index times out, but the index is created. Kibana then
complains that a .kibana index already exists, and another kibana instance must
be migrating despite it being the only kibana instance.